third_party/WebKit/LayoutTests/platform/linux/external/wpt/web-animations/animation-model/keyframe-effects/effect-value-transformed-distance-expected.txt

This is a testharness.js-based test.
PASS A step-start function on a keyframe affects the resulting style
PASS A steps(1, start) function on a keyframe affects the resulting style
PASS A steps(2, start) function on a keyframe affects the resulting style
PASS A step-end function on a keyframe affects the resulting style
PASS A steps(1) function on a keyframe affects the resulting style
PASS A steps(1, end) function on a keyframe affects the resulting style
PASS A steps(2, end) function on a keyframe affects the resulting style
PASS A frames function on a keyframe affects the resulting style
PASS A linear function on a keyframe affects the resulting style
FAIL A ease function on a keyframe affects the resulting style assert_approx_equals: The width should be approximately 109.47963055884654 at 1100ms expected 109.47963055884654 +/- 0.01 but got 109.312
PASS A ease-in function on a keyframe affects the resulting style
FAIL A ease-in-out function on a keyframe affects the resulting style assert_approx_equals: The width should be approximately 101.97224534417684 at 1100ms expected 101.97224534417684 +/- 0.01 but got 101.953
FAIL A ease-out function on a keyframe affects the resulting style assert_approx_equals: The width should be approximately 116.05721544376388 at 1100ms expected 116.05721544376388 +/- 0.01 but got 116.078
FAIL A easing function which produces values greater than 1 on a keyframe affects the resulting style assert_approx_equals: The width should be approximately 171.60874310990272 at 1100ms expected 171.60874310990272 +/- 0.01 but got 172.219
FAIL A easing function which produces values less than 1 on a keyframe affects the resulting style assert_approx_equals: The width should be approximately 77.13128857063907 at 1100ms expected 77.13128857063907 +/- 0.01 but got 76.9531
PASS Linear-equivalent cubic-bezier keyframe easing applied to an effect with a step-start function does not alter the result
PASS Linear-equivalent cubic-bezier keyframe easing applied to an effect with a steps(1, start) function does not alter the result
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a steps(2, start) function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 0ms expected "50.3438px" but got "50px"
PASS Linear-equivalent cubic-bezier keyframe easing applied to an effect with a step-end function does not alter the result
PASS Linear-equivalent cubic-bezier keyframe easing applied to an effect with a steps(1) function does not alter the result
PASS Linear-equivalent cubic-bezier keyframe easing applied to an effect with a steps(1, end) function does not alter the result
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a steps(2, end) function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 500ms expected "50.3438px" but got "50px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a frames function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "25.0938px" but got "25px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a linear function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "25.0938px" but got "25px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a ease function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "41.875px" but got "41.8594px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a ease-in function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "8.95312px" but got "9.32812px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a ease-in-out function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "12.9375px" but got "12.875px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a ease-out function does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "38.0469px" but got "37.9688px"
FAIL Linear-equivalent cubic-bezier keyframe easing applied to an effect with a easing function which produces values greater than 1 does not alter the result assert_equals: The 'width' of the animated elements should be equal at 250ms expected "100px" but got "102.406px"
PASS Linear-equivalent cubic-bezier keyframe easing applied to an effect with a easing function which produces values less than 1 does not alter the result
Harness: the test ran to completion.
